Represent percentage as a double (0.0 - 1.0) rather than an integer (0 - 100).

+++ src/KingForms/ApplicationActionProgress.cs
@@ -5,12 +5,12 @@ public sealed class ApplicationActionProgress
     public static ApplicationActionProgress None { get; } = new ApplicationActionProgress(x => { }, x => { });
 
     public IProgress<string> Text { get; }
-    public IProgress<int> Percent { get; }
+    public IProgress<double> Percent { get; }
 
-    public ApplicationActionProgress(Action<string> textHandler, Action<int> percentHandler)
-    : this(new Progress<string>(textHandler), new Progress<int>(percentHandler)) { }
+    public ApplicationActionProgress(Action<string> textHandler, Action<double> percentHandler)
+    : this(new Progress<string>(textHandler), new Progress<double>(percentHandler)) { }
 
-    public ApplicationActionProgress(IProgress<string> textProgress, IProgress<int> percentProgress)
+    public ApplicationActionProgress(IProgress<string> textProgress, IProgress<double> percentProgress)
     {
         Text = textProgress;
         Percent = percentProgress;
